requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pysentimiento==0.7.3
transformers==4.44.2
torch==2.3.1
scikit-learn==1.5.1
psycopg2-binary==2.9.9
